Intermediate Commit: Add base structure for check-in functionality
This commit gets up to letting users define some products and make a
request to the stripe create checkout session api, and then print that
api response. I want to implement logging and tracing so that I can see
those api request in the future, should the need arise. Thus, I'm making
this commit so that I can add the logging and tracing without worrying