Add tracing
Add the beginning of trace infrastructure. More instrumentation needs to
be added, but it was sufficient to help me develop the code to send
requests to the Stripe API and get users to the point where they can
make a payment on Stripe. Next, handle those payments being returned to
us and displaying a confirmation page.