Query Stripe for available products & prices
Instead of defining the products and prices at app instantiation,
dynamically load them by querying Stripe. Currently, the loaded prices
aren't actually used to update the AppState, that needs to be
implemented next. This feature lets the cashier log onto the Stripe
dashboard and change a price and add/remove a product there, rather than
having to interact with this app.