Check-in products now have `required_roles`
Check-in products have the requisite information to limit viewership to
users that contain certain roles. Next I need to implment the logic to
actually filter the viewership. This will be best done with a HashSet,
but currently the roles field on the users struct is not a hash set, to
make it easier to deserialize in the database. I'm going to try out
using hash sets to see how that works.