Redirect to check-in post-login
This is a temporary UX fix. Technically, I should return you to the page
you were on before you clicked the login button. That would take more
effort that I have time for right now, so I'm just going to do this
because the majority of people just need to see the check-in page.