Fix nav hamburger double click bug
Commit `8f9f6231` added a feature that clicking off of the navbar when
on small screens would close the navbar. This was useful, however it
introduced a bug that required users to double click the navbar to get
access to it again, which was annoying. I spent 2 hours fixing this tiny
bug. I am going to blame Jessica Rose Davidson for that.