working shopping cart
The shopping cart works. Right now it is a little bit tough to update
the quantity of items in the cart/remove items from the cart, but that
would require time/effort to fix and I don't really want to do that
right now. Next is to update the checkout sessions to handle the cart.